Drop template tag
