--- wagtail/admin/templates/wagtailadmin/icons/sprite.html
@@ -1,3 +1,0 @@
-{% load wagtailadmin_tags %}
-
-{% icons %}
